Potentially deflake integration tests
Reduce the likelihood of the 'echo' test failing due to the following
exception being thrown:

    > OSError: [Errno 113] No route to host

This was happening because the first process (in the background)
was not starrted before the command whose output is being asserted on
is run.

To work around this, add a sleep call prior to calling run to increase
the likelihood of the the backgroud process being started. This
introduces new timeout constants used only by this test in order to
account for this.

Also attempt to deflake "Larger packets: flow control". This test
requires that 2 background processes are started _after_ the candump
foreground process is started, since the candump process needs to
observe all the CAN communication.

Bump the delay prior to starting the two background processes and
increase the timeout on the candump command.